feat: let any logged-in user trigger their own sync (drops admin gate)
