+++ LetterboxdSync/ServiceRegistrator.cs
@@ -8,6 +8,7 @@ public class ServiceRegistrator : IPluginServiceRegistrator
 {
     public void RegisterServices(IServiceCollection serviceCollection, IServerApplicationHost applicationHost)
     {
+        serviceCollection.AddSingleton<LetterboxdSyncRunner>();
         serviceCollection.AddHostedService<PlaybackHandler>();
     }
 }
